COMPLETE_VIRAL_SYSTEM_PACKAGE/requirements.txt
# =================================================================
# VIRAL YOUTUBE DOCUMENTARY GENERATION SYSTEM - Dependencies
# =================================================================
# 
# This file lists ALL Python packages required for the system to work.
# Installation: pip install -r requirements.txt
#
# EXPLANATION FOR BEGINNERS:
# - Lines starting with # are comments (notes for humans, ignored by Python)
# - Each line is a package name with version (package==version)
# - Version numbers ensure compatibility (prevents breaking changes)
# =================================================================

# ===== CORE AI & LLM =====
# These handle the AI agents and language models

# Anthropic Claude - Our main AI engine
anthropic==0.34.0          # Official Anthropic SDK for Claude API

# LangChain - Framework for building AI agent workflows
langchain==0.2.16          # Main LangChain library
langchain-anthropic==0.1.23  # Anthropic integration for LangChain
langchain-community==0.2.16  # Community integrations
langchain-core==0.2.38      # Core abstractions
langgraph==0.2.5           # Graph-based workflows (multi-agent coordination)

# OpenAI - For embeddings (converting text to vectors)
openai==1.41.0             # OpenAI API client

# ===== VECTOR DATABASES =====
# For storing and searching context/knowledge

# ChromaDB - Local vector database (lightweight, easy)
chromadb==0.5.5            # Vector database for semantic search
chromadb-client==0.5.5     # Client library

# Qdrant - Alternative vector database (more powerful)
qdrant-client==1.11.1      # Client for Qdrant vector database

# ===== EMBEDDINGS & NLP =====
# For converting text into numerical representations

sentence-transformers==3.0.1  # Generate text embeddings
transformers==4.44.2          # Hugging Face transformers (NLP models)
torch==2.4.0                  # PyTorch (deep learning framework)
tiktoken==0.7.0               # Token counting for OpenAI models

# ===== WEB SCRAPING & APIs =====
# For fetching data from websites and APIs

requests==2.32.3           # HTTP requests (API calls)
beautifulsoup4==4.12.3     # HTML parsing (web scraping)
lxml==5.3.0                # XML/HTML parser
httpx==0.27.0              # Async HTTP client

# ===== ACADEMIC RESEARCH =====
# For searching academic databases

scholarly==1.7.11          # Google Scholar API
arxiv==2.1.3               # arXiv paper search
biopython==1.84            # PubMed integration

# ===== DATA PROCESSING =====
# For handling various data formats

python-dotenv==1.0.1       # Load environment variables from .env
pydantic==2.8.2            # Data validation and settings
pyyaml==6.0.2              # YAML file parsing (configs)
toml==0.10.2               # TOML file parsing

# ===== UTILITIES =====
# Helper libraries for various tasks

colorlog==6.8.2            # Colored console logging
tqdm==4.66.5               # Progress bars
typing-extensions==4.12.2  # Type hints for older Python versions

# ===== YOUTUBE & VIDEO =====
# For YouTube integration and video processing

yt-dlp==2024.8.6           # YouTube video downloader
pytube==15.0.0             # YouTube metadata and downloads
youtube-transcript-api==0.6.2  # Get YouTube transcripts

# ===== DATABASE & STORAGE =====
# For data persistence

sqlalchemy==2.0.32         # SQL database toolkit
alembic==1.13.2            # Database migrations

# ===== TESTING & DEVELOPMENT =====
# For testing and debugging

pytest==8.3.2              # Testing framework
pytest-asyncio==0.23.8     # Async testing
black==24.8.0              # Code formatter
flake8==7.1.1              # Code linter
mypy==1.11.2               # Type checker

# ===== OPTIONAL: JSTOR =====
# Uncomment if you have JSTOR API access
# jstor-api==0.1.0         # JSTOR academic database (requires institutional access)

# =================================================================
# INSTALLATION INSTRUCTIONS
# =================================================================
#
# STEP 1: Create virtual environment
#   python3 -m venv venv
#
# STEP 2: Activate virtual environment
#   # On Windows:
#   venv\Scripts\activate
#   
#   # On Mac/Linux:
#   source venv/bin/activate
#
# STEP 3: Install all dependencies
#   pip install -r requirements.txt
#
# STEP 4: Verify installation
#   python scripts/quick_test.py
#
# =================================================================
# TROUBLESHOOTING
# =================================================================
#
# ERROR: "Could not find a version that satisfies the requirement"
# FIX: Update pip first: pip install --upgrade pip
#
# ERROR: "Microsoft Visual C++ 14.0 is required" (Windows)
# FIX: Install Visual Studio Build Tools
#
# ERROR: "Failed building wheel for torch"
# FIX: Install PyTorch separately from pytorch.org
#
# ERROR: Import errors after installation
# FIX: Make sure virtual environment is activated
#
# =================================================================
